test_url/requirements.txt
drf_yasg
djangorestframework